add animation to welcome loading screen
